Only commit successful directory changes
If you try to cd into a directory that you don't have permissions to
open, the file browser would become blank with no way to recover since
we had already changed the file browser state before knowing if the
directory change would succeed.

This patch ensures that we only commit the directory change after we've
successfully read it.